Fix PeerManager mutex bottleneck
Do not try to acquire a MessageQueue mutex while holding the PeerManager mutex. A MessageQueue mutex may already be held for message processing and waiting for these will delay releasing the global PeerManager mutex. Instead, start goroutines to update the MessageQueue asynchronously so that the PeerManager lock can be released sooner.